build-suggestions/4.11: Raise minor_min to 4.10.16
We work really hard to test updates and identify regressions before
shipping releases.  But there is a chance we miss something, and need
to declare it as a conditional risk affecting 4.10 to 4.11 updates.
[1] landed web-console support for conditional updates in 4.10.16, so
raising the minumum here ensures all users will have that convenient
experience navigating any conditional risks which may have turned up
by the time we're recommending supported updates to 4.11.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=2087041#c5