blocked-edges/4.10.*: Declare ReleaseDataWithHyphenPrefix risk
Generated by writing 4.10.14-release-data-with-hyphen-prefix.yaml and
then distributing to the other impacted releases with:

  $ for Z in $(seq 15 18); do sed "s/4[.]10[.]14/4.10.${Z}/g" blocked-edges/4.10.14-release-data-with-hyphen-prefix.yaml > "blocked-edges/4.10.${Z}-release-data-with-hyphen-prefix.yaml"; done

The fix happened in 4.10.20 [1], but we don't need to declare
conditional risks for 4.10.19, because we never signed it, and
declaring risks for unsigned release images confused graph-data
presubmits [2]:

  Failed to find signatures for digest sha256:299ce2f9520282173c290b3e237aaa22526ab98ddc67604f2c6fc10e342f5940: [\n    \"Error fetching https://mirror.openshift.com/pub/openshift-v4/signatures/openshift/release/sha256=299ce2f9520282173c290b3e237aaa22526ab98ddc67604f2c6fc10e342f5940/signature-1 - 404 Not Found\", ...

where that is one of the architecture-specific 4.10.19 releases:

  $ oc adm release info quay.io/openshift-release-dev/ocp-release@sha256:299ce2f9520282173c290b3e237aaa22526ab98ddc67604f2c6fc10e342f5940 | head -n4
  Name:      4.10.19
  Digest:    sha256:299ce2f9520282173c290b3e237aaa22526ab98ddc67604f2c6fc10e342f5940
  Created:   2022-06-15T16:27:31Z
  OS/Arch:   linux/s390x

Because the releases that hash to hyphens depend on the architecture
[3], we could have declared two series of risks, one for 'to:
4.10.14+amd64' that looked at born-by-4.6 (which might have touched
4.6.64) and one for 'to: 4.10.14+s390x' that looked at born-by-4.9
(which might have touched 4.9.11).  That would have reduced the number
of clusters who matched the PromQL, and needed to manually evaluate
the risk to assess the impacted updates.  But because the current
ClusterVersion history pruning can remove the originally installed
version [4], and 4.6 was a while back, we could have clusters that:

1. Touched 4.6.64 and acquired hyphen-prefixed release data.
2. Accumulated enough subsequent updates to push 4.6 entries out of
   the history [4].
3. Are now serving cluster_version{type="initial"} with 4.7 or later
   versions.

To reduce that risk, and because we are interested in exposing
customers to the conditional risk interfaces to collect feedback, we
are using architecture-agnostic risk declarations and using <= 4.9
initial versions as matching rule broad enough to catch all clusters
which might be at risk.  Clusters where the risk matches can evaluate
their exposure as discussed in the linked solution, and decide to
accept the risk if they determine they are not vulnerable.

The risk is for everyone trying to leave 4.10.(14 <= z < 20), not just
the ones who got there from 4.10.z.  But I'm only declaring the
conditional risk for 4.10.z' -> 4.10.(14 <= z < 20), because only 4.10
cluster-version operators will have the conditional risk support.
4.9.z clusters updating into the impacted releases and then bumping
into the bug while trying to leave will have to follow the recommended
mitigation steps to get out.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=2094078#c6
[2]: https://prow.ci.openshift.org/view/gs/origin-ci-test/pr-logs/pull/openshift_cincinnati-graph-data/2118/pull-ci-openshift-cincinnati-graph-data-master-e2e/1541879405234622464#1:build-log.txt%3A24
[3]: https://bugzilla.redhat.com/show_bug.cgi?id=2091770#c17
[4]: https://issues.redhat.com/browse/OTA-664