blocked-edges/4.11.7-ovn-namespace: 4.11.7 does not fix the regression
Generated by rewording the 4.11.6 'message' to be less specific about
when the regression was introduced (and instead just point out that
the particular target version being considered is vulnerable).  An
then:

  $ sed 's/4[.]11[.]6/4.11.7/g' blocked-edges/4.11.6-ovn-namespace.yaml >blocked-edges/4.11.7-ovn-namespace.yaml

[1] is tracking the 4.11.z fix.

We could use an expression for 'from' to recommend 4.11.6 to 4.11.7
updates for OVN clusters, because they are already vulnerable to this
issue on 4.11.6, so the update to 4.11.7 doesn't make this worse.  For
a previous example, see 4d282f716d (blocked-edges/4.7.28: Restore
nuanced edge blocking, 2021-09-02, #1032).  But for now, I'm assuming
there will not be many OVN clusters on 4.11.6, and that those clusters
will be able to wait for 4.11.8 or whatever patch release actually
fixes [1], even though that means missing out on some 4.11.7 fixes in
the meantime.

[1]: https://issues.redhat.com//browse/OCPBUGS-1750