hack/show-edges: Add --root-version
To make it easier to subset the graph when looking at something
recent, without all the noise of older releases.  For example:

  $ hack/show-edges.py --root-version 4.10.42 stable-4.11
  4.10.42 -> 4.10.43
  4.10.42 -> 4.10.45
  4.10.42 -(blocked: ARM64SecCompError524)-> 4.11.17
  4.10.42 -(blocked: ARM64SecCompError524)-> 4.11.18
  4.10.42 -(blocked: ARM64SecCompError524)-> 4.11.20
  4.10.43 -> 4.10.45
  4.10.43 -(blocked: ARM64SecCompError524)-> 4.11.17
  4.10.43 -(blocked: ARM64SecCompError524)-> 4.11.18
  4.10.43 -(blocked: ARM64SecCompError524)-> 4.11.20
  4.10.45 -(blocked: ARM64SecCompError524)-> 4.11.20
  4.11.17 -> 4.11.18
  4.11.17 -> 4.11.20
  4.11.18 -> 4.11.20

We may want to pair this with --leaf-version or some such if we wanted
to to rule out some newer releases, but I'm leaving that off for now,
and we can revisit if we have a need.