Convert PromQL risks to Always risks in releases older than 8 weeks
This works around a CVO bug/design decision where we only evaluate one
newly enumerated risk every 10 minutes in an effort to avoid overwhelming
the monitoring stack with requests. However this creates a bad UX where
if there are many risks to evaluate in the set of available update paths
it could be N-1 * 10 minutes before the set of recommended updates are
computed.

This preserves the notification of issue while largely being a no-op
because clusters have, currently, had better update paths for at least
12 weeks.

We intend to fix the CVO bug, but that won't fix the issue in the
deployed fleet.

See: https://issues.redhat.com/browse/OCPBUGS-5469