hack/stabilization-changes: Dedup notifications
Avoid redundancies like:

  Cincinnati stabilization:
  * FAILED ARM64SecCompError524 affects 4.11.26.  Either declare a fix version or extend the risk to 4.11.27.
  * FAILED ARM64SecCompError524 affects 4.11.26.  Either declare a fix version or extend the risk to 4.11.27.
  * FAILED ARM64SecCompError524 affects 4.11.26.  Either declare a fix version or extend the risk to 4.11.27.
  * channels/fast-4.10: Promote 4.10.51. It was promoted...

when a release fails promotion into several channels for the same
reason (and we don't currently include the target channel in the
notification message, because it doesn't seem helpful).

With this change, we'll only get each message once.  And we'll
preserve the previous ordering, so specific channels are still grouped
together, regardless of the message strings the stabilize_channel
function returns.