Retire some older conditional edges which required PromQL eval
This is to work around the series of bugs where it takes 10mins per
`type: PromQL` risk to evaluate in versions prior to 4.12.9, 4.11.34,
and 4.10.56. If we reduce the number of outstanding PromQL risks by
converting them to Always risks when we've had superseding versions
out for quite some time we can improve the user experience.