Do not warn about risks fixed in earlier candidate to the same channel
When checking for concerns about risk extensions, take into account also
versions that are still candidates for the same channel. It is possible
that the (earlier) version that carries the information about the fix
(`fixedIn`) is also still just a candidate to the channel, and in that
case we would previously log a spurious error message.

For the case where 4.11.40 file contained `fixedIn: 4.11.41` and both
versions were still just candidates for stable:

https://github.com/openshift/cincinnati-graph-data/blob/e7567e45e4e1c4857f34f0cf22f341536c1e55af/blocked-edges/4.11.40-KeepalivedMulticastSkew.yaml#L3

Before:
t
```
* Recommend waiting to promote 4.11.40 to stable; it was promoted the feeder fast by ebf3b7c6fa (Merge pull request #3623 from openshift-ota-bot/promote-4.11.40-to-fast, 2023-05-18, 6 days, 8:35:15.178187)
* FAILED KeepalivedMulticastSkew affects 4.11.39.  Either declare a fix version or extend the risk to 4.11.41.
```

After:

```
* Recommend waiting to promote 4.11.40 to stable; it was promoted the feeder fast by ebf3b7c6fa (Merge pull request #3623 from openshift-ota-bot/promote-4.11.40-to-fast, 2023-05-18, 6 days, 8:38:59.720912)
* Recommend waiting to promote 4.11.41 to stable; it was promoted the feeder fast by 362a228041 (Merge pull request #3644 from openshift-ota-bot/promote-4.11.41-to-fast, 2023-05-24, 12:22:06.721565)
```