hack/exposure-length: Add a script showing known-exposure length
Timeline for an update-risk bug:

1. Releases named with vulnerable code.
2. Someone notices the issue.
3. Update risk declared.
4. Release with the fix named.
5. fixedIn added so that release can enter candidate-*.

The duration T2 - T1 can be long, because many of these risks are rare
corner cases.  T3 - T2 can be short, or it can be long if it takes us
some time to understand the severity of the issue.  This script
measures T5 - T3, so we have a convenient handle on that phase.