Remove KubeletRestartBug risk
While this is a definite regression there are no normally supported
operations which trigger a kubelet restart outside of a drain and reboot
cycle. Additionally, the risk is simply that pods get restarted which is
an expected pod lifecycle event which workloads must tolerate due to any
number of conditions such as eviction policies. Therefore we are removing
this risk, we hope that a fix for this defect will be available soon.

We recommend continuing to follow this bug for progress
https://issues.redhat.com/browse/OCPBUGS-14605