OWNERS_ALIASES: Add sre-approvers alias
This will allow us to grant SRE approval powers to declare
newly-discovered update risks even when all the
cincinnati-graph-data-approvers are asleep or otherwise not available.

Populated from the openshift-team-sd-sre-team-leads team [1].
Automatically update the membership with:

  $ GITHUB_TOKEN=#FIXME: https://docs.github.com/en/rest/overview/authenticating-to-the-rest-api
  $ CONTENT="$(grep -B1000 sre-approvers: OWNERS_ALIASES; curl -sL -H "Accept: application/vnd.github+json" -H "Authorization: Bearer ${GITHUB_TOKEN}" https://api.github.com/orgs/openshift/teams/openshift-team-sd-sre-team-leads/members | jq -r '.[].login' | grep -v openshift-ci-robot | sort | sed 's/^/    - /')"
  $ echo "${CONTENT}" >OWNERS_ALIASES

Where the grep assumes that 'sre-approvers:' is in the final
OWNERS_ALIASES line that we want to keep, and the API curl fills in
the subsequent portion with the current set of SD-SRE team leads.

[1]: https://github.com/orgs/openshift/teams/openshift-team-sd-sre-team-leads