blocked-edges/4.13.*-PersistentVolumeDiskIDSymlinks: Convert to Always
To mitigate slow cache warming [1], like 99161903ac (Retire some older
conditional edges which required PromQL eval, 2023-05-09, #3590).  The
fixedIn 4.13.4 entered stable-4.13 back in June:

  $ git --no-pager blame channels/stable-4.13.yaml | grep ' 4[.]13[.]4$'
  90eca13b2 (Automated Stabilization Robot 2023-06-29 08:03:38 +0000 40) - 4.13.4

so it's unlikely that there are all that many folks still interested
in updating to it, instead of a more recent 4.13.z.  And anyone who is
interested can still read the message and ticket, decide whether
they're exposed, and update if they feel they aren't exposed (or if
they feel they are exposed but can handle being bit).

Generated with:

  $ for X in blocked-edges/*PersistentVolumeDiskIDSymlinks*; do CONTENT="$(grep -B100 matchingRules: "${X}"; printf -- '- type: %s\n' Always)"; echo "${CONTENT}" > "${X}"; done

[1]: https://issues.redhat.com/browse/OTA-942