blocked-edges: Declare OVNWebhookUserConflict for 4.14.0-rc.3, rc.4, and 4.15.0-ec.0
The risk is for updating out of the impacted releases to hypothetical
future releases with the user change, e.g. from 4.14.0-rc.3 to 4.14.0.
But we're warning on * -> 4.14.0-rc.3 (and similar) to help folks
dodge the sticky updates entirely, because 4.14.0-rc.2 -> 4.14.0 are
not expected to ever have running-as-root webhook listeners.

Pattern-matching:

* apiserver_storage_objects from b573872208
 (blocked-edges/4.13.*-PerformanceProfilesCPUQuota: Declare new risk,
 2023-06-27, #3786),

* egressips.k8s.ovn.org from 894fab7974
  (blocked-edges/4.11.9-ovn-namespace: 4.11.9 does not fix the
  regression, 2022-10-12, #2628), and

* _id from 5cb2e93728 (blocked-edges/4.11.*-KeepalivedMulticastSkew:
  Explicit _id="", 2023-05-09, #3591).

Using cluster_installer with the hypershift invoker is new for this
commit, and in this case I'm using it to declare HyperShift clustres
not exposed to the risk.

Generated by manually writing the rc.e risk, and then copying it around with:

  $ for VERSION in 4.14.0-rc.4 4.15.0-ec.0; do sed "s/4.14.0-rc.3/${VERSION}/" blocked-edges/4.14.0-rc.3-OVNWebhookUserConflict.yaml > "blocked-edges/${VERSION}-OVNWebhookUserConflict.yaml"; done