blocked-edges/4.14.0: Drop updates from 4.13.17 and 18
We'd raised the floor for 4.13-to-4.14 in 1db9474be8a4 (4.14: Bump
`minor_min` to 4.13.19 to pickup the SCC gate, 2023-10-25, #4295), but
that only applies to automatically generated new releases.  ART
created the 4.14.0 metadata by hand, and included 4.13.17 and 4.13.18.
But we do not recommend folks update from those to 4.14 without
passing through 4.13.19 or later to get the SCC Upgradeable checker.
There are some trade offs between this commit's silent drop
vs. declaring an Always risk:

* A silent drop simplifies update graphs, not even presenting the
  not-recommended updates which could distract customers that don't
  care about those updates.
* A silent drop may mean we do not need to support customers who
  update from 4.13.17 or 18 directly to 4.14.0 and have some mutated
  SCCs stomped.  Or at least, there are not explicit docs one way or
  the other about whether customers who do this will be supported.
  And with the updates silently dropped, the number of customers who
  do this update is expected to be very low.
* An Always risk might have more customers thing "I probably didn't
  mutate my SCCs", accept the risk, and then be surprised when they
  actually had mutated their SCCs and the SCCs got stomped.
* An Always risk would reduce the chances that folks saw:

    $ oc adm release info -o json quay.io/openshift-release-dev/ocp-release:4.14.0-x86_64 | jq -r '.metadata.previous[]' | grep '^4[.]13[.]'
    4.13.17
    4.13.18
    4.13.19

  and then opened support cases about why they didn't see 4.14.0 as a
  direct-hop update target in their 4.13.17 or 18 cluster (the
  transparency issues that conditional update risks was designed to
  address).

Whether Always or silent-drops are better for customers is unclear.
But soon 4.14.1 will come out, and after that, folks caring about
updates to 4.14.0 will likely be very rare, so it doesn't seem like
it's worth pinning down a technological winner, and we're going with
silent-drop.