Add 4 days to the stable-4.11 channel delay
Currently, 4.11 and 4.12 are on an alternating weekly schedule. This means
that 4.11.z will go in one week but won't have an update path to 4.12 for
another roughly 7 days while the following 4.12 release cooks in fast
channels. By delaying things for another 4 days we're splitting the
difference and the maximum delay for those who aren't able to use channels
directly, aka ROSA clusters.