blocked-edges/4.14*: Declare AzureDefaultVMType
Generated by writing the 4.14.1 risk by hand, and then running:

  $ curl -s 'https://api.openshift.com/api/upgrades_info/graph?channel=candidate-4.14&arch=amd64' | jq -r '.nodes[] | .version' | grep '^4[.]14[.]' | grep -v '^4[.]14[.][01]$' | while read VERSION; do sed "s/4.14.1/${VERSION}/" blocked-edges/4.14.1-AzureDefaultVMType.yaml > "blocked-edges/${VERSION}-AzureDefaultVMType.yaml"; done

I also manually added the silent-drop to 4.14.0.  We have almost
entirely avoided silent drops since growing the ability to declare
conditional risks in 4.10.  But f0dc7e86d1 (blocked-edges/4.14.0: Drop
updates from 4.13.17 and 18, 2023-10-26, #4301) decided to use silent
drops for 4.13.17 and 4.13.18 to 4.14.0.  As described in that commit
message, there are trade-offs between silent-drops and an Always risk
for those updates.  With this commit, we double-down on the
silent-drop approach.

Because we're dropping 4.13.19 to 4.14.0 after it has been a supported
update for so long, there is a larger risk (than there was for 4.13.17
and 4.13.18 updates) of customers noticing the drop and being confused
about where the 4.13.19 to 4.14.0 update went.  That's why we
developed the conditional update system in the first place.

That risk is mitigated by the fact that 4.14.0 is fairly old by now,
with many subsequent 4.14.z that fix a number of other issues.  So we
do not expect there to be much residual interest in 4.13.19 to 4.14.0
updates.

If it turns out that there is enough "where did 4.13.19 to 4.14.0 go?"
support load to warrant a pivot, future work could move us to
explicitly-declared risk for all of the issues from 4.13.z to 4.14.0,
including the original "4.13.19 adds a guard you want first" that
f0dc7e86d1 was delivering.  But the price of that pivot is the:

  A silent drop may mean we do not need to support customers who
  update from 4.13.17 or 18 directly to 4.14.0 and have some mutated
  SCCs stomped...

trade-off discussed in f0dc7e86d1.