EWSECRLegacyCredProvider: Detect actual ECR usage from pullspecs
Joel Speed says:

> Looking into the fix for this, I found sources saying that ECR creds should be used for any image matching the following globs
> - "*.dkr.ecr.*.amazonaws.com"
> - "*.dkr.ecr.*.amazonaws.com.cn"
> - "*.dkr.ecr-fips.*.amazonaws.com"
> - "*.dkr.ecr.us-iso-east-1.c2s.ic.gov"
> - "*.dkr.ecr.us-isob-east-1.sc2s.sgov.gov"

We can use the `kube_pod_container_info` metric that is labeled with an `image` label that contains pullspec, so we can regex-match over these and detect that a cluster is actually using ECR from the presence of a pod containing a container using an image from ECR.

The regex is a bit hairy so I added a comment. [PromQL is using RE2 syntax](https://prometheus.io/docs/prometheus/latest/querying/basics/):

> All regular expressions in Prometheus use [RE2 syntax](https://github.com/google/re2/wiki/Syntax)

Here is the regex101.com session for toying with the regex: https://regex101.com/r/OIjQoS/1

I have used the PromQL conceptually on b01 and b03, testing for the
usage of a CoreDNS image coming from the CI registry:

```promql
topk(1,
  cluster_infrastructure_provider{_id="",type="AWS"}
  or
  0 * cluster_infrastructure_provider{_id=""}
)
* on () group_left()
(topk(1,
  group(kube_pod_container_info{_id="",image=~"registry[.]ci[.]openshift[.]org/ci/coredns:.*"})
  or on ()
  0 * group(kube_pod_container_info{_id=""})
))
```

Build01 uses it and I get a metric with value 1, Build03 does not and I get a metric with value 0.

I modified one of the YAML files manually and then I used my IDE's search-and-replace feature to modify all others.