blocked-edges/4.15.0-rc.0-GCPMintModeRoleAdmin: Declare risk
The PromQL logic is pretty similar to 4.13.9-AWSMintModeWithoutCredentials.yaml

The exposure is probably somewhere in the 4.15 ECs or RCs, but I'm
naively opening with a blanket 4.14 -> 4.15 to get this out the door.
We can refine the threshold later as we pin down exactly when in 4.15
the creds started using custom roles.

Generated by manually writing the 4.15.0-rc.0 risk, and copying out to
other 4.15 with:

  $ curl -s 'https://api.openshift.com/api/upgrades_info/graph?channel=candidate-4.15&arch=amd64' | jq -r '.nodes[] | .version' | grep '^4[.]15[.]' | grep -v '^4[.]15[.]0-rc[.]0$' | while read VERSION; do sed "s/4.15.0-rc.0/${VERSION}/" blocked-edges/4.15.0-rc.0-GCPMintModeRoleAdmin.yaml > "blocked-edges/${VERSION}-GCPMintModeRoleAdmin.yaml"; done