OTA-1179: OWNERS: Grant sre-approvers root approval status (#3955)
* OWNERS_ALIASES: Add sre-approvers alias

This will allow us to grant SRE approval powers to declare
newly-discovered update risks even when all the
cincinnati-graph-data-approvers are asleep or otherwise not available.

Populated from the openshift-team-sd-sre-team-leads team [1].
Automatically update the membership with:

  $ GITHUB_TOKEN=#FIXME: https://docs.github.com/en/rest/overview/authenticating-to-the-rest-api
  $ CONTENT="$(grep -B1000 sre-approvers: OWNERS_ALIASES; curl -sL -H "Accept: application/vnd.github+json" -H "Authorization: Bearer ${GITHUB_TOKEN}" https://api.github.com/orgs/openshift/teams/openshift-team-sd-sre-team-leads/members | jq -r '.[].login' | grep -v openshift-ci-robot | sort | sed 's/^/    - /')"
  $ echo "${CONTENT}" >OWNERS_ALIASES

Where the grep assumes that 'sre-approvers:' is in the final
OWNERS_ALIASES line that we want to keep, and the API curl fills in
the subsequent portion with the current set of SD-SRE team leads.

[1]: https://github.com/orgs/openshift/teams/openshift-team-sd-sre-team-leads

* OWNERS: Grant sre-approvers root approval status

cincinnati-graph-data-approvers are not on-call, and update risks that
impact the SRE-managed fleet may be discovered while no
cincinnati-graph-data-approvers are available to declare them.  This
commit grants root approval powers to sre-approvers, so they can:

* Update the sre-approvers alias in OWNERS_ALIASES, as described in
  9982225f97 (OWNERS_ALIASES: Add sre-approvers alias, 2023-08-08).
* Manage update risks in blocked-edges/.
* Use '/override ...' comments in pull requests (which require
  root-directly approval powers) if needed to bypass any flaking CI
  and land emergency update risk declarations.

I'm not including sre-approvers as reviewers, because when
cincinnati-graph-data-approvers are available, they'll handle review.
And when they are not available, sre-approvers will be driving an
emergency response using internal communication channels, and will not
need Prow/blunderbuss help in attracting reviewers.

I'm adding OWNERS to all the other child directories besides
blocked-edges, because site reliability engineers will not need access
to those locations.  By explicitly dropping them as approvers, Prow
advice about relevant approvers will avoid pulling in SREs who are not
interested in non-emergency repository activity.

* README: Add a section walking through a new managed-cluster risk

To make it easier for folks who declare risks rarely to declare a risk
that only affects managed clusters.  Once
cincinnati-graph-data-approvers wake up, we can round about further
PromQL refinement, whether to declare the risk for all clusters,
managed or not, and so forth.

* *channels/OWNERS: Drop these files for now

Avoid [1]:

  Verifying blocked edge files are valid
  Verifying channel files are valid
  Error: "/code/channels/OWNERS": is missing an extension

as graph-data.rs/src/verify_yaml.rs says "hey, this isn't a channel
file?".  Removing this file will allow SRE to approve changes within
the channels directory, but the risk of them approving something that
graph-data approvers would not have approved seems very low.

If/when we have confidence that all consuming Cincinnati (including
customer-deployed Cincinnati) have learned to ignore OWNERS when
loading directories, we can revert this commit.

[1]: https://prow.ci.openshift.org/view/gs/test-platform-results/pr-logs/pull/openshift_cincinnati-graph-data/3955/pull-ci-openshift-cincinnati-graph-data-master-e2e/1753098744288514048#1:build-log.txt%3A18-20